refactor(gates): split monolithic gates.ts into red/green/regression + helpers
