+++ .github/actions/gates/green/action.yml
@@ -8,5 +8,5 @@ runs:
   steps:
     - uses: sw2m/octoscript/action@main
     - id: run
-      run: (await import("./scripts/vsdd/promote/to-pr/gates.ts")).green();
+      run: (await import("./scripts/vsdd/promote/to-pr/green.ts")).green();
       shell: octoscript {0}
