+++ .github/actions/gates/regression/action.yml
@@ -8,5 +8,5 @@ runs:
   steps:
     - uses: sw2m/octoscript/action@main
     - id: run
-      run: (await import("./scripts/vsdd/promote/to-pr/gates.ts")).regression();
+      run: (await import("./scripts/vsdd/promote/to-pr/regression.ts")).regression();
       shell: octoscript {0}
